Add filter layers, radiation data, openaipi

+++ src/main/java/com/example/eco_map/persistence/model/City.java
@@ -25,6 +25,7 @@ import org.locationtech.jts.geom.Point;
 import org.locationtech.jts.geom.Polygon;
 import org.springframework.data.annotation.CreatedDate;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -37,8 +38,8 @@ import java.util.UUID;
                 columns = {
                         @ColumnResult(name = "id", type = UUID.class),
                         @ColumnResult(name = "name", type = String.class),
-                        @ColumnResult(name = "lat", type = Double.class),
-                        @ColumnResult(name = "lon", type = Double.class)
+                        @ColumnResult(name = "lat", type = BigDecimal.class),
+                        @ColumnResult(name = "lon", type = BigDecimal.class)
                 }
         )
 )
